Crowd_Detec-main/Crowd_Detec-main/requirements.txt
# AI Crowd Detection Model - Dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy==1.26.4
matplotlib>=3.5.0
pyyaml>=6.0
pandas>=1.3.0
pillow>=8.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0
fastapi>=0.70.0
uvicorn[standard]>=0.15.0